Change router directory structure Route handlers were moved to separate files so as to reduce clutter. Handlers were also named to ease testing.
